gcc/config/arm: make "decl_one_only" functions weak (ICE fix)
This commit is a workaround for the internal compiler error (ICE):

 error: comdat-local function called by XYZ outside its comdat

The problem appears to affect unified destructors ("*D4Ev"), which are
put in a COMDAT section, then get cloned to D0Ev, D1Ev and D2Ev, all
of these calling the original D4Ev.  The make_decl_one_only() function
made these "public", which however gets reverted on D4Ev by
maybe_thunk_body().  However, only public functions are allowed to be
called from another COMDAT, which leads to the ICE (and crashes gcc).

There is one way out of this: making D4Ev "weak", which is the way
chosen by the djgpp port (via an MAKE_DECL_ONE_ONLY override).  Tests
showed that this also works for my WinCE port.  As of now, this is a
workaround, because I don't understand the issue yet fully.